Migrating to New API for Grafana
Fixed #87

ChangeLog:
  - Added support for a new SDK: https://github.com/grafana/grafana-api-golang-client
  - Migrated All APIs to use the new SDK

BREAKING CHANGES:
  - SFTP support dropped.

+++ integration_tests/organizations_integration_test.go
@@ -14,7 +14,7 @@ func TestOrgs(t *testing.T) {
 	orgs := apiClient.ListOrganizations()
 	assert.Equal(t, len(orgs), 1)
 	mainOrg := orgs[0]
-	assert.Equal(t, mainOrg.ID, uint(1))
+	assert.Equal(t, mainOrg.ID, int64(1))
 	assert.Equal(t, mainOrg.Name, "Main Org.")
 
 }
